lib: pcm_mix: Moved PCM mix from application
Tests and code moved from applications: nrf5340_audio
