+++ .github/test-spec.yml
@@ -44,6 +44,7 @@
   - "boards/arm/nrf5340_audio_dk_nrf5340/**/*"
   - "tests/nrf5340_audio/**/*"
   - "tests/lib/contin_array/**/*"
+  - "tests/lib/pcm_mix/**/*"
 
 "CI-iot-samples-test":
   - "include/modem/at_monitor.h"
