+++ CODEOWNERS
@@ -116,6 +116,7 @@ Kconfig*                                  @tejlmand
 /lib/modem_jwt/                           @jayteemo @SeppoTakalo
 /lib/modem_attest_token/                  @jayteemo
 /lib/qos/                                 @simensrostad
+/lib/pcm_mix/                             @koffes @alexsven @erikrobstad @rick1082 @gWacey
 /lib/contin_array/                        @koffes @alexsven @erikrobstad @rick1082 @gWacey
 /modules/                                 @tejlmand
 /modules/mcuboot/                         @de-nordic @nordicjm
