+++ applications/nrf5340_audio/src/utils/CMakeLists.txt
@@ -12,5 +12,4 @@ target_sources(app PRIVATE
 	       ${CMAKE_CURRENT_SOURCE_DIR}/pcm_stream_channel_modifier.c
 	       ${CMAKE_CURRENT_SOURCE_DIR}/tone.c
 	       ${CMAKE_CURRENT_SOURCE_DIR}/uicr.c
-		   ${CMAKE_CURRENT_SOURCE_DIR}/pcm_mix.c
 )
