+++ lib/CMakeLists.txt
@@ -41,3 +41,4 @@ add_subdirectory_ifdef(CONFIG_QOS qos)
 add_subdirectory_ifdef(CONFIG_IDENTITY_KEY identity_key)
 add_subdirectory_ifdef(CONFIG_SFLOAT sfloat)
 add_subdirectory_ifdef(CONFIG_CONTIN_ARRAY contin_array)
+add_subdirectory_ifdef(CONFIG_PCM_MIX pcm_mix)
