+++ lib/pcm_mix/CMakeLists.txt
@@ -1,0 +1,12 @@
+#
+# Copyright (c) 2022 Nordic Semiconductor
+#
+# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+#
+
+zephyr_library()
+zephyr_library_sources(
+	pcm_mix.c
+)
+
+zephyr_include_directories(.)
