+++ tests/nrf5340_audio/pcm_stream_channel_modifier/Kconfig
@@ -1,10 +1,6 @@
 # Copyright (c) 2022 Nordic Semiconductor ASA
 # SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
 
-module = PCM_MIX
-module-str = pcm-mix
-source "subsys/logging/Kconfig.template.log_config"
-
 module = PSCM
 module-str = pscm
 source "subsys/logging/Kconfig.template.log_config"
